Correção de bugs

+++ src/main/kotlin/br/com/steam/dto/GameReviewDTO.kt
@@ -2,9 +2,9 @@ package br.com.steam.dto
 
 data class GameReviewDTO(
         val datePosted: String? = null,
-        val negative: Int? = null,
-        val positive: Int? = null,
-        val hoursPlayed: Int? = null,
+        val negative: Long? = null,
+        val positive: Long? = null,
+        val hoursPlayed: Long? = null,
         val recommendation: String? = null,
         val review: String? = null
 )
